update request complete

+++ assignment/public/js/customjs.js
@@ -1,14 +1,17 @@
 // custom JavaScript
 
 const deleteTask = async (id) => {
-  let taskID = Number(id);
-  await fetch(`/${taskID}`, {
-    method: "DELETE",
-    headers: {
-      "Content-type": "application/json",
-    },
-  });
+  const confirmation = confirm("You really want to delete this task...?");
+  if (confirmation) {
+    let taskID = Number(id);
+    await fetch(`/${taskID}`, {
+      method: "DELETE",
+      headers: {
+        "Content-type": "application/json",
+      },
+    });
 
-  const card = document.getElementById(id);
-  card.remove();
+    const card = document.getElementById(id);
+    card.remove();
+  }
 };
